requirements.txt
Django
djangorestframework
pandas
SQLAlchemy
markdown
django-filter
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
youtube-easy-api
numpy
num2word
textblob
nltk
newsapi
newsapi-python
praw
scikit-learn
schedule
--upgrade git+https://github.com/twintproject/twint.git@origin/master#egg=twint

django-admin-list-filter-dropdown
pyspellchecker
langdetect

pytz
twint
spellchecker
APScheduler